feat: Improve tests workflow
