+++ .gitignore
@@ -2,3 +2,5 @@ target
 /my-o7-app
 
 node_modules
+
+features.txt
